move assets to its own file
