+++ applications/bci_visualization/operators/reconstruction/convert_to_voxels_operator.py
@@ -12,7 +12,7 @@ import cupy as cp
 import numpy as np
 from numpy.typing import NDArray
 
-from processing.reconstruction import ExtinctionCoefficient, HbO
+from processing.reconstruction.hbo import ExtinctionCoefficient, HbO
 from holoscan.core import (
     ExecutionContext,
     InputContext,
